+++ src/components/atom/shape-action/ShapeAction.js
@@ -40,8 +40,8 @@ const ShapeAction = ({
 ShapeAction.propTypes = {
   children: PropTypes.instanceOf(Object),
   isButton: PropTypes.bool,
-  type: PropTypes.oneOf(['primary', 'success', 'danger']),
-  fill: PropTypes.oneOf(['primary', 'success', 'danger']),
+  type: PropTypes.oneOf(['primary', 'success', 'danger', 'gray-light']),
+  fill: PropTypes.oneOf(['primary', 'success', 'danger', 'gray-light']),
   shape: PropTypes.oneOf(['squared', 'rounded']),
   size: PropTypes.oneOf([1, 2, 3, 4]),
   isLink: PropTypes.bool